requirements.txt
PyYAML
bottle
